+++ apps/api/src/marketplace/marketplace.module.ts
@@ -14,6 +14,7 @@ import { MarketplaceLawncareService } from './entities/marketplace-lawncare-serv
 import { MarketplaceLawncarePackage } from './entities/marketplace-lawncare-package.entity';
 import { MarketplaceLawncarePackageSubscription } from './entities/marketplace-lawncare-package-subscription.entity';
 import { MarketplaceLawncarePropertyProfile } from './entities/marketplace-lawncare-property-profile.entity';
+import { MarketplaceHouseCleaningPropertyProfile } from './entities/marketplace-house-cleaning-property-profile.entity';
 import { MarketplacePestService } from './entities/marketplace-pest-service.entity';
 import { MarketplacePestPackage } from './entities/marketplace-pest-package.entity';
 import { MarketplacePestPackageSubscription } from './entities/marketplace-pest-package-subscription.entity';
@@ -32,6 +33,7 @@ import { ServiceRequestsModule } from '../service-requests/service-requests.modu
       MarketplaceAddOn, MarketplaceFrequencyDiscount, MarketplaceSubscription,
       MarketplaceSubscriptionEvent, MarketplaceLawncareService, MarketplaceLawncarePackage,
       MarketplaceLawncarePackageSubscription, MarketplaceLawncarePropertyProfile,
+      MarketplaceHouseCleaningPropertyProfile,
       MarketplacePestService, MarketplacePestPackage, MarketplacePestPackageSubscription, MarketplacePestPropertyProfile,
       VendorCapability, ServicePrice,
     ]),